Move part construction to `AggregatePartClass`
