+++ server/src/main/java/io/spine/server/aggregate/AggregatePartClass.java
@@ -22,12 +22,14 @@ package io.spine.server.aggregate;
 
 import io.spine.annotation.Internal;
 import io.spine.server.BoundedContext;
+import io.spine.server.entity.AbstractEntity;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.server.aggregate.AggregatePart.GenericParameter.AGGREGATE_ROOT;
+import static java.lang.String.format;
 
 /**
  * Provides type information on an aggregate part class.
@@ -60,10 +62,76 @@ public final class AggregatePartClass<A extends AggregatePart> extends Aggregate
         return rootClass;
     }
 
+    /**
+     * Creates a new {@code AggregatePart} entity and sets it to the default state.
+     *
+     * @param ctor the constructor to use
+     * @param <I>  the type of entity IDs
+     * @param <A>  the type of the entity
+     * @return an {@code AggregatePart} instance
+     */
+    static <I, A extends AbstractEntity<I, ?>> A create(Constructor<A> ctor,
+                                                        AggregateRoot<I> root) {
+        checkNotNull(ctor);
+        checkNotNull(root);
+
+        try {
+            final A result = ctor.newInstance(root);
+            return result;
+        } catch (InvocationTargetException | InstantiationException |
+                 IllegalAccessException | IllegalArgumentException e) {
+            throw new IllegalStateException(e);
+        }
+    }
+
+    /**
+     * Obtains the constructor for the passed aggregate part class.
+     *
+     * <p>The part class must have a constructor with ID and {@code AggregateRoot} parameters.
+     *
+     * <p>Returns the constructor if the first parameter is aggregate ID
+     * and the second constructor parameter is subtype of the {@code AggregateRoot}
+     * For example:
+     * <pre>{@code
+     *
+     * // A user-defined AggregateRoot:
+     * class CustomAggregateRoot extends AggregateRoot{...}
+     *
+     * // An AggregatePart for the CustomAggregateRoot:
+     * class CustomAggregatePart extends AggregatePart<...>{
+     *
+     *     // The expected constructor:
+     *     CustomAggregatePart(AnAggregateId id, CustomAggregateRoot root){...}
+     *     }
+     * }
+     * </pre>
+     *
+     * <p>Throws {@code IllegalStateException} in other cases.
+     *
+     * @param cls the {@code AggregatePart} class
+     * @param <A> the {@code AggregatePart} type
+     * @param <I> the ID type
+     * @return the constructor
+     * @throws IllegalStateException if the entity class does not have the required constructor
+     */
+    static <A extends AggregatePart<I, ?, ?, R>, I, R extends AggregateRoot<I>>
+    Constructor<A> getConstructor(Class<A> cls) {
+        checkNotNull(cls);
+
+        final Class<R> aggregateRootClass = getRootClass(cls);
+        try {
+            final Constructor<A> ctor = cls.getDeclaredConstructor(aggregateRootClass);
+            ctor.setAccessible(true);
+            return ctor;
+        } catch (NoSuchMethodException ignored) {
+            throw noSuchConstructor(cls, aggregateRootClass);
+        }
+    }
+
     @Override
     protected Constructor<A> findConstructor() {
         @SuppressWarnings("unchecked") // The cast is protected by generic params.
-        final Constructor<A> ctor = AggregatePart.getConstructor(value());
+        final Constructor<A> ctor = getConstructor(value());
         return ctor;
     }
 
@@ -93,5 +161,15 @@ public final class AggregatePartClass<A extends AggregatePart> extends Aggregate
             throw new IllegalStateException(e);
         }
         return result;
+    }
+
+    private static IllegalStateException noSuchConstructor(Class<?> aggregatePartClass,
+                                                           Class<?> aggregateRootClass) {
+        final String errMsg =
+                format("%s class must declare a constructor with %s parameter type.",
+                       aggregatePartClass.getName(),
+                       aggregateRootClass.getName());
+        final NoSuchMethodException cause = new NoSuchMethodException(errMsg);
+        throw new IllegalStateException(cause);
     }
 }
